fmd-1127 - Tidy Up
  - move single funciton in file to where it is used
  - delete graphql_helpers
  - delete unused functions
